* commit '1050a583c7955640f7026af28ed6203b5998ab8c': (66 commits)
  Preparation for release 1.6.0.
  [ITB-299] Corrected the theme issues (image alignment) on FF and IE.
  [ITB-269] The conformance statement creation wizard is now streamlined;
  [ITB-291] Conformance statements and results now correctly reflect the actor's role in the test cases.
  [ITB-170] Variables now support only name and type attributes.
  [ITB-382] XML Match validator implementation.
  [ITB-383] Processing transactions are now optional
  [ITB-311] Processing handler to create text tokens
  [ITB-384] Processing steps now get logged
  [ITB-379, ITB-320] Dropdown list implementation for user requests and reset button.
  [ITB-280] Step results are now added to the STEP_SUCCESS map.
  [ITB-279, ITB-376, ITB-378] Improvements to the handling of list types
  [ITB-377] State is no longer reloaded upon error where not explicitly allowed.
  [ITB-295] Exception stacktraces and messages reported to users are now filtered to hide detailed technical info.
  [ITB-282] Defined explicit ordering for test cases in test suites
  [ITB-304] CSV imports now normalise whitespace and escape characters that could be used for formula injection attacks.
  Allowed the DB password to be set via environment property.
  [ITB-347, ITB-348, ITB-349, ITB-351, ITB-353] Improvement of session tracking
  [ITB-370] Virus scanning is now added to all user uploads.
  [ITB-374] Corrected the display of the password fields.
  ...

+++ gitb-ui/app/controllers/RepositoryService.scala
@@ -3,63 +3,74 @@ package controllers
 import java.io._
 import java.nio.file.{Files, Paths}
 
-import javax.xml.transform.stream.StreamSource
 import com.gitb.tbs.TestStepStatus
 import com.gitb.tpl.TestCase
 import com.gitb.utils.XMLUtils
-import controllers.util.{ParameterExtractor, Parameters, ResponseConstructor}
+import config.Configurations
+import controllers.util.{AuthorizedAction, ParameterExtractor, Parameters, ResponseConstructor}
+import exceptions.ErrorCodes
 import javax.inject.Inject
+import javax.xml.transform.stream.StreamSource
 import managers._
 import models.ConformanceCertificate
+import org.apache.commons.codec.binary.Base64
 import org.apache.commons.codec.net.URLCodec
 import org.apache.commons.io.FileUtils
 import org.apache.commons.lang3.StringUtils
 import org.slf4j.LoggerFactory
 import play.api.mvc._
-import utils.{JacksonUtil, JsonUtil, MimeUtil, RepositoryUtils}
+import utils._
 
 /**
  * Created by serbay on 10/16/14.
  */
-class RepositoryService @Inject() (testCaseManager: TestCaseManager, testSuiteManager: TestSuiteManager, reportManager: ReportManager, testResultManager: TestResultManager, conformanceManager: ConformanceManager, specificationManager: SpecificationManager) extends Controller {
+class RepositoryService @Inject() (testCaseManager: TestCaseManager, testSuiteManager: TestSuiteManager, reportManager: ReportManager, testResultManager: TestResultManager, conformanceManager: ConformanceManager, specificationManager: SpecificationManager, authorizationManager: AuthorizationManager) extends Controller {
 	private val logger = LoggerFactory.getLogger(classOf[RepositoryService])
 	private val codec = new URLCodec()
 
   private val TESTCASE_STEP_REPORT_NAME = "step.pdf"
 
-	def getTestSuiteResource(testId: String, filePath:String): Action[AnyContent] = Action {
-		implicit request =>
-      val testCase = testCaseManager.getTestCaseForIdWrapper(testId).get
-      val testSuite = testSuiteManager.getTestSuiteOfTestCaseWrapper(testCase.id)
-      var filePathToLookup = codec.decode(filePath)
-      if (!filePath.startsWith(testSuite.shortname) && !filePath.startsWith("/"+testSuite.shortname)) {
-        // Prefix with the test suite name.
-        filePathToLookup = testSuite.shortname + "/" + filePathToLookup
-      }
-      val file = RepositoryUtils.getTestSuitesResource(specificationManager.getSpecificationById(testCase.targetSpec), filePathToLookup)
-			logger.debug("Reading test resource ["+codec.decode(filePath)+"] definition from the file ["+file+"]")
-			if(file.exists()) {
-				Ok.sendFile(file, true)
-			} else {
-				NotFound
-			}
+	def getTestSuiteResource(testId: String, filePath:String) = AuthorizedAction { request =>
+    authorizationManager.canViewTestSuiteResource(request, testId)
+    val testCase = testCaseManager.getTestCaseForIdWrapper(testId).get
+    val testSuite = testSuiteManager.getTestSuiteOfTestCaseWrapper(testCase.id)
+    var filePathToLookup = codec.decode(filePath)
+    if (!filePath.startsWith(testSuite.shortname) && !filePath.startsWith("/"+testSuite.shortname)) {
+      // Prefix with the test suite name.
+      filePathToLookup = testSuite.shortname + "/" + filePathToLookup
+    }
+    filePathToLookup = StringUtils.replaceOnce(filePathToLookup, testSuite.shortname, testSuite.filename)
+    // Ensure that the requested resource is within the test suite folder (to avoid path traversal)
+    val spec = specificationManager.getSpecificationById(testSuite.specification)
+    val testSuiteFolder = RepositoryUtils.getTestSuitesResource(spec, testSuite.filename)
+    val file = RepositoryUtils.getTestSuitesResource(spec, filePathToLookup)
+    logger.debug("Reading test resource ["+codec.decode(filePath)+"] definition from the file ["+file+"]")
+    if (file.exists() && file.toPath.normalize().startsWith(testSuiteFolder.toPath.normalize())) {
+      Ok.sendFile(file, true)
+    } else {
+      NotFound
+    }
 	}
 
-	def getTestStepReport(sessionId: String, reportPath: String): Action[AnyContent] = Action { implicit request=>
+	def getTestStepReport(sessionId: String, reportPath: String) = AuthorizedAction { request =>
 //    34888315-6781-4d74-a677-8f9001a02cb8/4.xml
+    authorizationManager.canViewTestResultForSession(request, sessionId)
+
 		val sessionFolder = reportManager.getPathForTestSessionWrapper(sessionId, true).toFile
     var path: String = null
-    if (reportPath.startsWith(sessionId)) {
+    path = reportPath
+      .replace("__SQS__", "[")
+      .replace("__SQE__", "]")
+
+    if (path.startsWith(sessionId)) {
       // Backwards compatibility.
-      val pathParts = StringUtils.split(codec.decode(reportPath), "/")
+      val pathParts = StringUtils.split(codec.decode(path), "/")
       path = pathParts(1)
     } else {
-      path = reportPath
+      path = path
     }
     path = codec.decode(path)
     val file = new File(sessionFolder, path)
-
-    logger.debug("Reading test step report ["+codec.decode(reportPath)+"] from the file ["+file+"]")
 
 		if(file.exists()) {
       //read file incto a string
@@ -76,8 +87,9 @@ class RepositoryService @Inject() (testCaseManager: TestCaseManager, testSuiteMa
 		}
 	}
 
-  def exportTestStepReport(sessionId: String, reportPath: String): Action[AnyContent] = Action { implicit request=>
+  def exportTestStepReport(sessionId: String, reportPath: String) = AuthorizedAction { request =>
     //    34888315-6781-4d74-a677-8f9001a02cb8/4.xml
+    authorizationManager.canViewTestResultForSession(request, sessionId)
     var path: String = null
     if (reportPath.startsWith(sessionId)) {
       // Backwards compatibility.
@@ -94,18 +106,21 @@ class RepositoryService @Inject() (testCaseManager: TestCaseManager, testSuiteMa
     if (!pdf.exists()) {
       if (file.exists()) {
         reportManager.generateTestStepReport(file.toPath, pdf.toPath)
-      } else {
-        NotFound
       }
     }
-    Ok.sendFile(
-      content = pdf,
-      fileName = _ => TESTCASE_STEP_REPORT_NAME
-    )
+    if (!pdf.exists()) {
+      NotFound
+    } else {
+      Ok.sendFile(
+        content = pdf,
+        fileName = _ => TESTCASE_STEP_REPORT_NAME
+      )
+    }
   }
 
-  def exportTestCaseReport(): Action[AnyContent] = Action.apply { implicit request =>
+  def exportTestCaseReport() = AuthorizedAction { request =>
     val session = ParameterExtractor.requiredQueryParameter(request, Parameters.SESSION_ID)
+    authorizationManager.canViewTestResultForSession(request, session)
     val testCaseId = ParameterExtractor.requiredQueryParameter(request, Parameters.TEST_ID)
 
     val folder = reportManager.getPathForTestSessionWrapper(codec.decode(session), true).toFile
@@ -138,13 +153,10 @@ class RepositoryService @Inject() (testCaseManager: TestCaseManager, testSuiteMa
     }
   }
 
-  def exportTestCaseReports(): Action[AnyContent] = Action { implicit request =>
-    NotFound
-  }
-
-  def exportConformanceStatementReport(): Action[AnyContent] = Action.apply { implicit request =>
-    val actorId = ParameterExtractor.requiredQueryParameter(request, Parameters.ACTOR_ID)
+  def exportConformanceStatementReport() = AuthorizedAction { request =>
     val systemId = ParameterExtractor.requiredQueryParameter(request, Parameters.SYSTEM_ID)
+    authorizationManager.canViewConformanceStatementReport(request, systemId)
+    val actorId = ParameterExtractor.requiredQueryParameter(request, Parameters.ACTOR_ID)
     val includeTests = ParameterExtractor.requiredQueryParameter(request, Parameters.TESTS).toBoolean
     val reportPath = Paths.get(
       ReportManager.getTempFolderPath().toFile.getAbsolutePath,
@@ -167,10 +179,11 @@ class RepositoryService @Inject() (testCaseManager: TestCaseManager, testSuiteMa
     )
   }
 
-  def exportConformanceCertificateReport(): Action[AnyContent] = Action.apply { implicit request =>
-    val actorId = ParameterExtractor.requiredBodyParameter(request, Parameters.ACTOR_ID).toLong
+  def exportConformanceCertificateReport() = AuthorizedAction { request =>
     val systemId = ParameterExtractor.requiredBodyParameter(request, Parameters.SYSTEM_ID).toLong
     val communityId = ParameterExtractor.requiredBodyParameter(request, Parameters.COMMUNITY_ID).toLong
+    val actorId = ParameterExtractor.requiredBodyParameter(request, Parameters.ACTOR_ID).toLong
+    authorizationManager.canViewConformanceCertificateReport(request, communityId)
 
     val jsSettings = ParameterExtractor.requiredBodyParameter(request, Parameters.SETTINGS)
     var settings = JsonUtil.parseJsConformanceCertificateSettings(jsSettings, communityId)
@@ -204,49 +217,53 @@ class RepositoryService @Inject() (testCaseManager: TestCaseManager, testSuiteMa
     )
   }
 
-  def exportDemoConformanceCertificateReport(communityId: Long): Action[AnyContent] = Action.apply { implicit request =>
+  def exportDemoConformanceCertificateReport(communityId: Long) = AuthorizedAction { request =>
+    authorizationManager.canViewConformanceCertificateReport(request, communityId)
     val jsSettings = ParameterExtractor.requiredBodyParameter(request, Parameters.SETTINGS)
     var settings = JsonUtil.parseJsConformanceCertificateSettings(jsSettings, communityId)
-    val reportPath = Paths.get(
-      ReportManager.getTempFolderPath().toFile.getAbsolutePath,
-      "conformance_reports",
-      "c"+communityId,
-      "report_"+System.currentTimeMillis+".pdf"
-    )
-    try {
-      FileUtils.deleteDirectory(reportPath.toFile.getParentFile)
-    } catch  {
-      case e:Exception => {
-        // Ignore these are anyway deleted every hour
+
+    var response: Result = null
+    if (settings.keystoreFile.isDefined && Configurations.ANTIVIRUS_SERVER_ENABLED) {
+      val virusScanner = new ClamAVClient(Configurations.ANTIVIRUS_SERVER_HOST, Configurations.ANTIVIRUS_SERVER_PORT, Configurations.ANTIVIRUS_SERVER_TIMEOUT)
+      val scanResult = virusScanner.scan(Base64.decodeBase64(MimeUtil.getBase64FromDataURL(settings.keystoreFile.get)))
+      if (!ClamAVClient.isCleanReply(scanResult)) {
+        response = ResponseConstructor.constructBadRequestResponse(ErrorCodes.VIRUS_FOUND, "Keystore file failed virus scan.")
       }
     }
-    if (settings.includeSignature && (settings.keystorePassword.isEmpty || settings.keyPassword.isEmpty)) {
-      // The passwords need to be looked up from the stored data.
-      val storedSettings = conformanceManager.getConformanceCertificateSettingsWrapper(communityId)
-      val completeSettings = new ConformanceCertificate(settings)
-      completeSettings.keyPassword = Some(MimeUtil.decryptString(storedSettings.get.keyPassword.get))
-      completeSettings.keystorePassword = Some(MimeUtil.decryptString(storedSettings.get.keystorePassword.get))
-      settings = completeSettings.toCaseObject
+    if (response == null) {
+      val reportPath = Paths.get(
+        ReportManager.getTempFolderPath().toFile.getAbsolutePath,
+        "conformance_reports",
+        "c"+communityId,
+        "report_"+System.currentTimeMillis+".pdf"
+      )
+      try {
+        FileUtils.deleteDirectory(reportPath.toFile.getParentFile)
+      } catch  {
+        case e:Exception => {
+          // Ignore these are anyway deleted every hour
+        }
+      }
+      if (settings.includeSignature && (settings.keystorePassword.isEmpty || settings.keyPassword.isEmpty)) {
+        // The passwords need to be looked up from the stored data.
+        val storedSettings = conformanceManager.getConformanceCertificateSettingsWrapper(communityId)
+        val completeSettings = new ConformanceCertificate(settings)
+        completeSettings.keyPassword = Some(MimeUtil.decryptString(storedSettings.get.keyPassword.get))
+        completeSettings.keystorePassword = Some(MimeUtil.decryptString(storedSettings.get.keystorePassword.get))
+        settings = completeSettings.toCaseObject
+      }
+      reportManager.generateDemoConformanceCertificate(reportPath, settings)
+      response = Ok.sendFile(
+        content = reportPath.toFile,
+        fileName = _ => reportPath.toFile.getName
+      )
     }
-    reportManager.generateDemoConformanceCertificate(reportPath, settings)
-    Ok.sendFile(
-      content = reportPath.toFile,
-      fileName = _ => reportPath.toFile.getName
-    )
+    response
   }
 
-	def getTestCase(testId:String) = Action.apply { implicit request =>
-		val tc = testCaseManager.getTestCase(testId)
-    if (tc.isDefined) {
-      val json = JsonUtil.jsTestCase(tc.get).toString()
-      ResponseConstructor.constructJsonResponse(json)
-    } else {
-      NotFound
-    }
-	}
-
-	def getTestCaseDefinition(testId: String) = Action.apply { implicit request =>
-		val tc = testCaseManager.getTestCase(testId)
+	def getTestCaseDefinition(testId: String) = AuthorizedAction { request =>
+    authorizationManager.canViewTestCase(request, testId)
+    val tc = testCaseManager.getTestCase(testId)
     if (tc.isDefined) {
       val file = RepositoryUtils.getTestSuitesResource(specificationManager.getSpecificationById(tc.get.targetSpec), tc.get.path)
       logger.debug("Reading test case ["+testId+"] definition from the file ["+file+"]")
@@ -260,15 +277,25 @@ class RepositoryService @Inject() (testCaseManager: TestCaseManager, testSuiteMa
     }
 	}
 
-  def getTestCases() = Action.apply { request =>
-    val testCaseIds = ParameterExtractor.extractLongIdsQueryParameter(request)
-
-    val testCases = testCaseManager.getTestCases(testCaseIds)
+  def getAllTestCases() = AuthorizedAction { request =>
+    authorizationManager.canViewAllTestCases(request)
+    val testCases = testCaseManager.getAllTestCases()
     val json = JsonUtil.jsTestCasesList(testCases).toString()
     ResponseConstructor.constructJsonResponse(json)
   }
 
-  private def removeFile(file:File) = {
-    file.delete()
+  def getTestCasesForSystem(systemId: Long) = AuthorizedAction { request =>
+    authorizationManager.canViewTestCasesBySystemId(request, systemId)
+    val testCases = testCaseManager.getTestCasesForSystem(systemId)
+    val json = JsonUtil.jsTestCasesList(testCases).toString()
+    ResponseConstructor.constructJsonResponse(json)
   }
+
+  def getTestCasesForCommunity(communityId: Long) = AuthorizedAction { request =>
+    authorizationManager.canViewTestCasesByCommunityId(request, communityId)
+    val testCases = testCaseManager.getTestCasesForCommunity(communityId)
+    val json = JsonUtil.jsTestCasesList(testCases).toString()
+    ResponseConstructor.constructJsonResponse(json)
+  }
+
 }
